// wardn/frontend/lib/api/generated/model/mCPRegistryServerListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPRegistryListMetadata } from './mCPRegistryListMetadata';
import type { MCPRegistryServerResponse } from './mCPRegistryServerResponse';

export interface MCPRegistryServerListResponse {
  metadata: MCPRegistryListMetadata;
  servers: MCPRegistryServerResponse[];
}