// wardn/frontend/lib/api/generated/model/mCPRegistryServerResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { MCPRegistryResponseMeta } from './mCPRegistryResponseMeta';
import type { MCPServerDocument } from './mCPServerDocument';

export interface MCPRegistryServerResponse {
  _meta: MCPRegistryResponseMeta;
  server: MCPServerDocument;
}